GWTII-154: change the objects that are exchanged; add full options print panel

+++ plugin/print/impl/src/test/java/org/geomajas/plugin/print/client/widget/PrintWidgetMockStart.java
@@ -133,7 +133,7 @@ public abstract class PrintWidgetMockStart {
 		MapComponentInfo mapComponent = (MapComponentInfo) getChildOfType(pageComponentInfo.getChildren(), MapComponentInfo.class);
 		Assert.assertNotNull(mapComponent);
 		Assert.assertEquals(viewData.getApplicationId(), mapComponent.getApplicationId());
-		Assert.assertEquals(viewData.getRasterDpi(), (int) mapComponent.getRasterResolution());
+		Assert.assertEquals((int) viewData.getRasterDpi(), (int) mapComponent.getRasterResolution());
 
 		// check scalebar
 		ScaleBarComponentInfo scaleBarComponent = (ScaleBarComponentInfo)
